move instance fix to imputation_main
